Boom roasted, added sessions.
Added support for sessions.  There are three new config options,
sessionEnabled, sessionTimeout in minutes (leave blank for sessions
which last until browser is closed), and cookieKey.  The cookie key is
used to encrypt the session cookie.  Session data is all stored in the
cookie, so there is a limit of ~3KB on data stored in a session, if
more data needs to be stored it should be stored in a db.  Sessions are
initialized  at the start of the request and is included in reqData,
you can call reqData.session.set or reqData.session.get to get and set
session data.