adds bitchen new routing
See the routerTest.js file for a couple of examples, basically routes
can be any regular expression and you can use the $# syntax to do
replacement.  Matches are done starting with the last route and moving
up, once a match is found it stops.  It will match on the entire string
because it adds ^ and $ to the regex before it tries to match, this
means that a pattern like '/' will not match any route only exactly '/'.